feat: populate manifest column stats for rewritten data files
iceberg_rewrite_data_files previously wrote manifest entries with only
record count, file size, and the partition tuple. The COPY that produces
the compacted parquet files already runs with RETURN_STATS, but those
per-column statistics were discarded, so rewritten files had no
lower/upper bounds, value/null counts, or column sizes. Query-time file
pruning (and Iceberg metrics such as write.metadata.metrics.column.*)
therefore stopped working for any file produced by a rewrite.

Extract the RETURN_STATS -> IcebergDataFile stats population from
IcebergInsertGlobalState::AddFiles into a shared helper
(PopulateDataFileColumnStatsFromReturnStats) and use it from both INSERT
and the rewrite operator. This guarantees rewritten files carry the same
bounds/metrics as freshly inserted files, respecting per-column metrics
modes, variant bounds, and geometry bounding boxes. Rewritten files also
inherit the table sort order id when one is configured.

Add a catalog-agnostic test asserting that rewritten files retain column
bounds and that file pruning still prunes non-matching files.
